+++ IO/source/MshReader/MshReader2D.cpp
@@ -149,12 +149,12 @@ void MshReader2D::addElements() {
 
 void MshReader2D::defineBoundaryVertices() {
 	for (auto boundary = this->gridData->boundaries.begin(); boundary < this->gridData->boundaries.end(); boundary++) {
-		std::set<int> vertices;
+		std::set<long int> vertices;
 		for (auto j = boundary->lineConnectivity.cbegin(); j != boundary->lineConnectivity.cend(); j++) {
 			for (auto k = j->cbegin(); k != j->cend(); k++) {
-				vertices.insert(*k);
+				vertices.insert(static_cast<long int>(*k));
 			}
 		}
-		boundary->vertices = std::vector<int>(vertices.begin(), vertices.end());
+		boundary->vertices = std::vector<long int>(vertices.begin(), vertices.end());
 	}
 }
